Refactor of ToolBar::initTheme for maintainability
This is my toe in the water.

I rolled up the two unrolled loops to keep DRY and reduce LOC. I also inverted the nature of if statements to escape early instead to reduce the depth of indentation especially as XML can be very demanding on this.

I believe part of the original style was to match the C++ nesting in flow control to match the same in the XML structure. Although I have undone that I think that this version is easily to work with.

Note: These changes have not been fully tested.